backend/requirements.txt
# Core FastAPI dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
python-multipart>=0.0.6

# Database & ORM
sqlalchemy>=2.0.23
psycopg[binary]>=3.1.0
alembic>=1.13.0
pgvector>=0.2.0

# Authentication
python-jose[cryptography]>=3.3.0
firebase-admin>=6.8.0

# HTTP Client & AI
httpx>=0.25.0
langchain-core>=0.1.0
openai>=1.3.0
google-generativeai>=0.8.3

# Data Validation
pydantic[email]>=2.5.0

# File Processing
aiofiles>=23.2.1

# Optional: Development tools
# pytest>=7.4.0
# pytest-asyncio>=0.21.0
# black>=23.0.0
# isort>=5.12.0
# mypy>=1.7.0
